+++ NEWS.md
@@ -5,6 +5,7 @@
 * methods for OTU clustering
 * some reusable regular expressions
 * some new functions to work with multiple seqtab objects (`seqtab_stats_bind` and `seqtab_stats_plot`)
+* better feature naming options for `ft` objects
 
 # jakR2 0.2